refactor(car, test): remove util package
util package 삭제, test 문구 추가
