+++ src/main/java/com/wootechcamp/precourse/racingcar/Cars.java
@@ -1,7 +1,5 @@
 package com.wootechcamp.precourse.racingcar;
 
-import com.wootechcamp.precourse.racingcar.util.RandomNumberGenerator;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
